requirements.txt
gradio>=4.0
pydantic>=2.0
matplotlib>=3.7
numpy>=1.24
python-dotenv>=1.0
